add usePage hook
